Refactor index.d.ts imports
